Add files via upload
Our analysis focused on examining key aspects of the dataset to gain insights into Netflix's content. Firstly, we sought to determine whether Netflix had a higher number of movies or TV series. 
Additionally, we investigated the presence of older content in the dataset, specifically identifying when it was added and the nature of that content.
Furthermore, we aimed to identify the year with the highest and lowest content releases, allowing us to understand the trends and patterns of Netflix's content distribution over time. Additionally, we explored the release patterns within each month to ascertain which day of the month Netflix typically releases the most content.